api/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
filelock==3.15.4
fsspec==2024.6.1
idna==3.7
Jinja2==3.1.4
kombu==5.4.0
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.4.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.1
openai-whisper==20231117
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
redis==5.0.8
regex==2024.7.24
requests==2.32.3
setuptools==72.2.0
six==1.16.0
sqlparse==0.5.1
sympy==1.13.2
tiktoken==0.7.0
torch==2.4.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
